Add extra options for HTML output.
Changes along the lines of pull request 3 by github:andrewshadoura.

Add to limit the file size. If size is exceeded new file is created.

*Primitive* support of incremental updates: check the first message in
the first file if it is the same - overwrite, otherwise - append.
NB foolproof implementation would need to scan all files in sequence,
before deciding to append.